refactor: move auth verification to app.js
