+++ src/components/Homepage/HomePage.js
@@ -2,21 +2,8 @@ import Banner from "./Banner";
 import Feed from "./Feed";
 import PopularTags from "./PopularTags";
 import Header from "../Header/Header";
-import { useEffect } from "react";
-import { getAuthUser, resetInitial } from '../../feature/auth/authSlice';
-import { useDispatch, useSelector } from "react-redux";
 
 function HomePage() {
-  const dispatch = useDispatch();
-  const { user } = useSelector(state => state.auth);
-  useEffect(() => {
-    if (sessionStorage.getItem("token") && !user) {
-      dispatch(getAuthUser());
-    }                                                                                                                 
-    else if(!sessionStorage.getItem("token") && user){
-      dispatch(resetInitial());
-    }
-  }, [])
   return (
     <div>
       <Header />
